Extracted rule into abstract class for reuse
